Move InstallationCandidate to pip.index
Move InstallationCandidate from `pip.req.req_requirement` to
`pip.index`, the one place where it's used.
